@@ -1,21 +1,26 @@
 {
   "private": true,
+  "type": "module",
   "license": "(CC-BY-NC-4.0 OR Apache-2.0)",
   "author": "Brian Holt <btholt+next-course-starter@gmail.com>",
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "build": "next build && npm run csv",
     "export": "next build && next export",
-    "start": "next start"
+    "start": "next start",
+    "csv": "node csv/index.js"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.1.1",
     "gray-matter": "^4.0.3",
-    "highlight.js": "^11.4.0",
-    "marked": "^4.0.9",
-    "next": "^12.0.7",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "highlight.js": "^11.5.1",
+    "marked": "^4.0.16",
+    "next": "^12.1.6",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "title-case": "^3.0.3"
+  },
+  "devDependencies": {
+    "convert-array-to-csv": "^2.0.0"
   }
 }
