@@ -1,10 +1,13 @@
-const course = require("./course.json");
+import { readFileSync } from "fs";
+import path from "path";
+
+const buffer = readFileSync(path.join(process.cwd(), "./course.json"));
+const course = JSON.parse(buffer);
 const BASE_URL = course?.productionBaseUrl || "";
 
-module.exports = {
+const config = {
   basePath: BASE_URL,
   env: {
-    ROOT: __dirname,
     BASE_URL,
   },
   async redirects() {
@@ -21,3 +24,5 @@ module.exports = {
     return [];
   },
 };
+
+export default config;
